Se agrega las pruebas Unitarias Probadas de  Factura, Pedido, Item y Pago. Se modificaron las variables de las interfaces, las implementaciones y las pruebas unitarias de Factura, Pedido, Item y Pago a  tipo primitivo.

+++ Codigo/TiendaEnLineaJPAEclipseLink/src/main/java/edu/co/sena/instrumusic/modelo/jpa/dao/implementacion/FacturaDAOImpl.java
@@ -134,7 +134,7 @@ public class FacturaDAOImpl implements IFacturaDAO{
     }
 
     @Override
-    public List<Factura> findByTotal(double total) {
+    public List<Factura> findByTotal(float total) {
         
         EntityManager em = getEntityManager();
         List<Factura> facturaT = null;
